refactor(category): Moved persistense directory into infrastructure
