+++ module/category/src/Infrastructure/Persistence/Projector/DbalCategoryCreatedEventProjector.php
@@ -7,18 +7,18 @@
 
 declare(strict_types = 1);
 
-namespace Ergonode\Category\Persistence\Dbal\Projector;
+namespace Ergonode\Category\Infrastructure\Persistence\Projector;
 
 use Doctrine\DBAL\Connection;
-use Ergonode\SharedKernel\Domain\Aggregate\AttributeId;
 use Ergonode\Attribute\Domain\ValueObject\AttributeCode;
 use Ergonode\Category\Domain\Event\CategoryCreatedEvent;
+use Ergonode\SharedKernel\Domain\Aggregate\AttributeId;
 use JMS\Serializer\SerializerInterface;
 use Ramsey\Uuid\Uuid;
 
 /**
  */
-class CategoryCreatedEventProjector
+class DbalCategoryCreatedEventProjector
 {
     private const TABLE = 'category';
     private const VALUE_TABLE = 'attribute_value';
