+++ module/category/src/Infrastructure/Persistence/Query/DbalCategoryQuery.php
@@ -7,16 +7,16 @@
 
 declare(strict_types = 1);
 
-namespace Ergonode\Category\Persistence\Dbal\Query;
+namespace Ergonode\Category\Infrastructure\Persistence\Query;
 
 use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\Query\QueryBuilder;
-use Ergonode\SharedKernel\Domain\Aggregate\CategoryId;
 use Ergonode\Category\Domain\Query\CategoryQueryInterface;
+use Ergonode\Category\Domain\ValueObject\CategoryCode;
 use Ergonode\Core\Domain\ValueObject\Language;
 use Ergonode\Grid\DataSetInterface;
 use Ergonode\Grid\DbalDataSet;
-use Ergonode\Category\Domain\ValueObject\CategoryCode;
+use Ergonode\SharedKernel\Domain\Aggregate\CategoryId;
 
 /**
  */
